Mocks: For a variable returning `AnyObserver` type, implement non-failing default return value handling.
Summary:
For a `var value: AnyObserver<T> { get }` variable, the reasonable default value handling, in case the `get` handler is not provided by the test suite, would be to return an instance of `AnyObserver` invoking (an optional) `eventHandler` closure.
The unit test could then choose to act on new events sent to the observer; the default implementation would do nothing by default.