Add support for generating mocks for protocols containing generic methods.
Summary:
To build mock class for a protocol that has generic method requirements, e.g.,
```
    func buildPopoverPresenter<T>() -> AnyPopoverPresentable<T>
```
it should be able to provide handler for such methods.
The handler could use `Any` in place of generic types, e.g.:
```
    var buildPopoverPresenterHandler: (() -> AnyPopoverPresentable<Any>)? = nil
```
or use generic type declared at class level:
```
class Mock<_T> {
    var buildPopoverPresenterHandler: (() -> AnyPopoverPresentable<_T>)? = nil
}
```

In either case, the result of the invocation of the handler will have to be force-casted to the expected return type of the generic method.
In case the function takes generic parameters, these need to be force-casted to the internal type first.
The problem with the latter is that Sourcery/SourceKit doesn't really parse generic types for method parameters,
so use another Source annotation, `annotatedGenericTypes` to provide a hint of where the generic parameter is actually located.