Improve short method name generation, also generate full names of method mocks in case duplicates have the same number of parameters.
Summary:

1) Improve treatment of `argumentLabel` in `mockedMethodName` when `useShortName == false`:
produce `updateCropHandles` instead of `updateCropHandlesCropHandles` for such function:
```
    func update(cropHandles: [CGPoint])
```

2) When several functions would produce the same short name, there was already logic in place to generate long function names.
There was also an optimization to keep the short name for the shortest function, to keep the mock class shorter, e.g., for these two functions:
```
    func update(_ handle: CGPoint)
    func update(cropHandles: [CGPoint])
```
the first one would keep using the sort `update` form, and the second would get full `updateCropHandles` name.

However, when two functions have the same number of arguments, it would make sense to just use the long form for all of them:
```
    func update(cropRect: CGRect)
    func update(cropHandles: [CGPoint])
```
will result in `updateCropRect` and `updateCropHandles`, which improves consistency of mock generation.