Better support `Single` Rx return types.
Summary:
`Single` return type is mocked using `PublishSubject`, but returning `subject.asSingle()` is not enough to mimic the desired behaviour.
The desired behaviour is that when (in the unit test), a subject is sent a `.next` event, the Single would emit `.success` to its subscribers.
This commit adds the desired behaviour.