Change how generic type constraints are annotated with `generictype`: Use "where" clause syntax, which refers to each constraint separately in a comma-delimited string. BREAKING CHANGE.
Summary:
## 1 ##
Previously, generic types that have multiple constraints were annotated like this:
```
    // sourcery: genericType = "T: RawRepresentable, Hashable"
    func presentErrorPopover<T>(
        // sourcery: genericTypePlaceholder = "AnyErrorPopoverPresentable<{T}>"
        _ popoverPresenter: AnyErrorPopoverPresentable<T>) -> Observable<T> where T: RawRepresentable, T: Hashable

```

This notation is inconsistent with how constraints are written in "where" clause, and does not allow to reference several types.
Update the notation, so that each comma-separated part represents a distinct constraint.
The above declaration now needs to be written as (which is a **breaking change**):
```
    // sourcery: genericType = "T: RawRepresentable, T: Hashable"
```

Also, distinguish the first part of the constraint type so that constraints on subtypes can be created like this:
```
    /// sourcery: genericType = "S: Sequence, S.Iterator.Element: Object"
    func add<S>(
        // sourcery: annotatedGenericTypes = "{S}"
        _ objects: S,
        update: Bool) where S: Sequence, S.Iterator.Element: Object
```

## 2 ##
Another **breaking change** is that now multiple generic types defined in the type (in both `associatedtype` type annotation and `generictype` function annotations) are merged on the top level.
Now it's possible to have two functions with the same generic type name. Usually that's the intention anyway. The constraints from these declarations will be merged and applied to the one generic type.
```
    /// sourcery: genericType = "S: Sequence, S.Iterator.Element: Object"
    func add<S>(
        // sourcery: annotatedGenericTypes = "{S}"
        _ objects: S,
        update: Bool) where S: Sequence, S.Iterator.Element: Object

    /// sourcery: genericType = "S: Sequence, S.Iterator.Element: Object"
    func delete<S>(
        // sourcery: annotatedGenericTypes = "{S}"
        _ objects: S) where S: Sequence, S.Iterator.Element: Object
```

Generic types and constraints from the above declarations will be merged:
```
class RealmInteropMock<_Element, _S>: RealmInterop where _Element: Object, _S.Iterator.Element: Object, _S: Sequence {
    typealias Element = _Element
    typealias KeyType = _KeyType
}
```

If that's not the intention, it's still possible to name generic types differently.