Add the ability to override mock method names with `sourcery: methodName = <METHOD_NAME>` annotation.
Summary:
In some cases, it's hard to resolve duplicate function names from a protocol definition in a meaningful way, e.g.:
```
    func delete(_ object: Object)
    func delete<S>(_ objects: S) where S: Sequence, S.Iterator.Element: Object
    func delete<Element>(_ objects: List<Element>) where Element: Object
    func delete<Element>(_ objects: Results<Element>) where Element: Object
```

Last 3 functions only differ by argument type names, which is hard to take into account in the mock generator, because usually we want to generate the shortest possible names for function handlers to make mock class readable and easy to use.
For such complex situations, let's use an annotation so that a human can help the computer to help the human.